web_labeler/requirements.txt
fastapi
uvicorn[standard]
numpy
opencv-python
Pillow
ultralytics
